Add files via upload
This api created in Java Srcipt and json packages, allows you as a user of a company X of computer products to be able to see, modify, filter by Id, update or delete the products from the database which is made with the tool from MongoDB. To be able to do all these actions you must first be able to register with a username, email and password (they can be any), and then do a SignIn and this will provide you with a token, this token is valid for one day and it is unique per user. Likewise, roles such as "admin" or "moderator" can be assigned and with these they can be assigned to different routes to make the database more private, although the default role. Also, the code has a function for pagination, it is only required to have the "mongoose-paginate-v2" library.

Each product in the database has an ID, a name, a category, a description (if the user wants to put it), its creation date and last modification date.